[CI Visibility] Split CODEOWNERS responsibilities
